feat: support for WebDAV
