+++ src/main/kotlin/app/termora/sync/SyncConfig.kt
@@ -4,6 +4,7 @@ enum class SyncType {
     GitLab,
     GitHub,
     Gitee,
+    WebDAV,
 }
 
 enum class SyncRange {
