+++ src/main/resources/i18n/messages.properties
@@ -83,6 +83,7 @@ termora.settings.sync.last-sync-time=Last sync time
 termora.settings.sync.gist=Gist
 termora.settings.sync.token=Token
 termora.settings.sync.type=Type
+termora.settings.sync.webdav.help=WebDav storage address, e.g. https://yourhost/webdav/termora.json
 
 termora.settings.about=About
 termora.settings.about.author=Author
