+++ src/main/resources/i18n/messages_zh_CN.properties
@@ -85,6 +85,7 @@ termora.settings.sync.import.successful=导入数据成功
 termora.settings.sync.gist=片段
 termora.settings.sync.token=令牌
 termora.settings.sync.type=类型
+termora.settings.sync.webdav.help=WebDav 的存储地址，例如：https://yourhost/webdav/termora.json
 
 termora.settings.about=关于
 termora.settings.about.author=作者
