+++ src/main/resources/i18n/messages_zh_TW.properties
@@ -89,6 +89,7 @@ termora.settings.sync.import.successful=導入資料成功
 termora.settings.sync.gist=片段
 termora.settings.sync.token=令牌
 termora.settings.sync.type=類型
+termora.settings.sync.webdav.help=WebDav 的儲存位址，例如：https://yourhost/webdav/termora.json
 
 termora.settings.about=關於
 termora.settings.about.author=作者
